Implemented asynchronous mutator API
-C++ client done (Future and TableMutatorAsync, along with tests)
-ThriftAPI changes implemented
-Thrift tests updated

+++ src/cc/Hypertable/Lib/TableMutatorAsyncHandler.h
@@ -37,7 +37,7 @@ namespace Hypertable {
     /**
      * Constructor.  Initializes state.
      */
-    TableMutatorAsyncHandler(TableMutatorAsync *mutator, size_t scatter_buffer);
+    TableMutatorAsyncHandler(TableMutatorAsync *mutator, uint32_t scatter_buffer);
 
     /**
      */
@@ -45,7 +45,7 @@ namespace Hypertable {
 
   private:
     TableMutatorAsync *m_mutator;
-    size_t m_scatter_buffer;
+    uint32_t m_scatter_buffer;
   };
 }
